[probes.starlark] Split print_metric tests into starlark_metric_test.go
starlark_test.go was approaching 1300 lines. Move all print_metric
tests plus the metricsAfterRun / findMetric helpers into a dedicated
file. Same package; helpers stay accessible if other test groups need
them. starlark_test.go drops the metrics + payloadconfigpb + metricspb
imports that were only used by the moved block.
